解锁 SampleSmoke 2 skip：修复 IL 产物 TypeLoad System.Console 根因 + 整型提升重载打分
- 根因定位：MetadataReader.FindMethod 在合并 references 首命中 SDK 壳（System.Core.Managed.dll 的 Console facade 含 BCL 缺失的窄型重载）→ WriteLine<Byte> MemberRef scope 指向 System.Core.Managed → netcore/netfx 直跑 TypeLoad System.Console（ByteArrays 块，HelloWorld 正常）
- 修 1：SDK Console.co 移除 BCL 无对应的窄型重载（WriteLine/Write × u8/i8/i16/u16，8 个）——BCL 无 byte/short WriteLine（提升 int），SDK 对齐 BCL
- 修 2：ResolveOverloadByScore 整型提升打分（Cocoa+CSharp 两方言）：窄整型（i8/u8/i16/u16）→ i32 权重 0（唯一最优，对齐 C# byte→int 优先），其余隐式数值 1——否则移除窄型重载后 WriteLine(u8) 在 i32/i64/u32/u64 歧义
- SampleSmokeTests Tutorial_DotNet/NetFx_AllBlocks 移除 Skip（2→0 skip）；歧义诊断测试改为真正同分（i32→i64/f64）
- 全量 42040 pass / 0 skip / 0 fail

+++ src/Cocoa.Cs/Cocoa.Tests/CodeAnalysis/NamespaceOverloadTests.cs
@@ -177,9 +177,9 @@ function Main(): i32
         public void Evaluator_Overload_Ambiguous_ReportsDiagnostic()
         {
             var text = @"
-                function G(a: i32): i32
+                function G(a: i64): i32
                 {
-                    return a
+                    return i32(a)
                 }
 
                 function G(a: f64): i32
@@ -189,7 +189,7 @@ function Main(): i32
 
                 function Main(): i32
                 {
-                    return [G](u8(3))
+                    return [G](i32(3))
                 }
             ";
